requirements.txt
streamlit
pandas
pillow
pyzbar